requirements.txt
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
pdfplumber==0.11.5
python_docx==1.1.2
streamlit==1.41.1
torch==2.5.1
transformers==4.48.0
